Refactor team menu layout and add new components
