+++ src/components/teams/TeamMenuTitle.tsx
@@ -6,13 +6,15 @@ const TeamMenuTitle = ({
   subtitle: string;
 }) => {
   return (
-    <div className="flex flex-col gap-4 pt-4">
+    <div className="flex flex-col gap-4">
       <span>
-        <h3 className="text-xl font-semibold leading-7">{title}</h3>
-        <span className="text-sm text-gray-500 font-light">{subtitle}</span>
+        <h3 className="text-3xl font-semibold leading-7 text-center">
+          {title}
+        </h3>
+        <span className="text-sm text-gray-500 font-light text-center block mt-4">
+          {subtitle}
+        </span>
       </span>
-
-      <div className="w-full border-t border-gray-300" />
     </div>
   );
 };
