+++ src/components/teams/form/TeamAPIKey.tsx
@@ -66,7 +66,7 @@ export default function TeamAPIKeyServer({ team }: Props) {
               }}
             >
               <div className="px-3 py-2 flex justify-between items-center rounded-md shadow-sm ring-1 ring-inset ring-gray-300 focus-within:ring-2 focus-within:ring-inset focus-within:ring-indigo-600 sm:max-w-md">
-                <span className=" overflow-hidden text-ellipsis block select-none items-center text-gray-500 sm:text-base">
+                <span className=" overflow-hidden text-ellipsis block select-none items-center text-gray-500 sm:text-base max-w-[calc(100%-2rem)]">
                   {displayedKey}
                 </span>
