+++ src/components/teams/form/settings/TeamIntegrations.tsx
@@ -3,6 +3,7 @@ import TypefullyPanel from '../TypefullyPanel';
 import TeamAPIKey from '../TeamAPIKey';
 import { Team } from '@prisma/client';
 import TeamMenuTitle from '../../TeamMenuTitle';
+import TeamMenuContent from '../../TeamMenuContent';
 
 const TeamIntegrations = ({ team }: { team: Team }) => {
   return (
@@ -11,9 +12,11 @@ const TeamIntegrations = ({ team }: { team: Team }) => {
         title="Integrations"
         subtitle="  Increase your digest reach by integrating with other tools."
       />
-      <SlackPanel team={team} />
-      <TypefullyPanel team={team} />
-      <TeamAPIKey team={team} />
+      <TeamMenuContent>
+        <SlackPanel team={team} />
+        <TypefullyPanel team={team} />
+        <TeamAPIKey team={team} />
+      </TeamMenuContent>
     </div>
   );
 };
